init v8.0.0

+++ components/csi/csi2/include/drv/common.h
@@ -1,5 +1,19 @@
-/*
- * Copyright (C) 2017-2020 Alibaba Group Holding Limited
+ /*
+ * Copyright (C) 2017-2024 Alibaba Group Holding Limited
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 /******************************************************************************
@@ -101,8 +115,8 @@ typedef struct {
 
 struct csi_dev {
     unsigned long reg_base;
-    uint8_t       irq_num;
-    uint8_t       idx;
+    uint16_t      irq_num;
+    uint16_t      idx;
     uint16_t      dev_tag;
     void          (*irq_handler)(void *);
     void          (*irq_handler2)(uint32_t irqn, void *arg);
@@ -119,8 +133,8 @@ struct csi_dev {
 
 typedef struct {
     unsigned long    reg_base;
-    uint8_t          irq_num;
-    uint8_t          idx;
+    uint16_t         irq_num;
+    uint16_t         idx;
     uint16_t         dev_tag;
 } csi_perip_info_t;
