+++ components/mpegts/src/ts_misc.c
@@ -1,6 +1,20 @@
 /*
-* Copyright (C) 2018-2023 Alibaba Group Holding Limited
-*/
+ * Copyright (C) 2017-2024 Alibaba Group Holding Limited
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the License); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an AS IS BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 #ifndef TSMISC_HPP
 #define TSMISC_HPP
 #include <mpegts/ts_misc.h>
@@ -71,7 +85,7 @@ uint32_t ts_crc32(const uint8_t *data, size_t len)
  */
 uint64_t ts_read_pts(const uint8_t data[5])
 {
-    uint32_t val = 0;
+    uint64_t val = 0;
     uint64_t pts = 0;
 
     val = data[0];
